CrOS Settings: Create InputsSection class
Create a dedicated class for inputs and related settings. This class
holds all related load time data and search tags, and removes some
code duplication. When the OsSettingsRevampWayfinding feature is
enabled, input settings are parented to the Device section. When the
feature is disabled, they belong to the Languages section.

Bug: b:303111314, b:298100014, chromium:1113611
Test: Manually test input settings search results
Test: browser_tests --gtest_filter="OSSettings*"
Test: unit_tests --gtest_filter="DeviceSectionTest.*"
Change-Id: I0446aed5c9dffc709bae43231aaf7fe37f278a1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4908420
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206104}