Fix dangling pointer in DlpScopedFileAccessDelegate
This CL changes the delegate so that it fetches the dlp client
every time it is needed. It also adds a check to ensure that the client
is both available and alive. This avoids holding onto the global
address in the delegate class.

This also adds an indirection for fetching the the global singleton.
This should make it much easier to swap between the real impl and the
fake impl during tests without changing the runtime behaviour.

Fixed: 1462026
Change-Id: If1e0ce5e8872a430ec0a62ef4b3829888ffe34ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4903098
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Fred Shih <ffred@chromium.org>
Reviewed-by: Daniel Brinkers <brinky@google.com>
Cr-Commit-Position: refs/heads/main@{#1206131}