[omnibox][ml][chrome://omnibox] Use `Signals` struct in mojom (5/6)
Currently, signals are passed as an array of <key, value> pairs from the
cpp to the ts code. This CL changes it to a struct.

This is in preparation for a followup CL that will add an 'ml
calculator' where the user can manually tweak signals and view the ml
score. That will require communicating signals back from ts to cpp code.
Using <key, value> pairs to do so would have a high risk of typos and
extra places we'd have to keep the consistent when
adding/renaming/removing signals.

Refactor only; no user behavior change.

Bug: 1405555
Change-Id: I57f5b814066cc9b3c1496b16d5e7f16fde70220a
Low-Coverage-Reason: HARD_TO_TEST Modifying a webUI page handler method which is integrating with cpp infra and ts code. TRIVIAL_CHANGE Mostly existing code, this CL's just replacing an array with a struct.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907764
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Jun Zou <junzou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206147}