[CloudOpenMetrics] Add MetricState
Add the `MetricState` enum which tracks the state of a given metric.
Update the state in the `Metric::Log()` function.

The state is initialised to `kCorrectlyNotLogged` and becomes
`kCorrectlyLogged` or `kIncorrectlyLoggedMultipleTimes` in
`Metric::Log()`. In the latter case an ERROR is logged.

The other states can be reached later in the `CloudOpenMetrics`
destructor.

Bug: b/300861997
Change-Id: I092f43356b00e4b4ad2591b898cd8e5e67bb004c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4914779
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206173}