Change the calculation of Prerender.Experimental.ReceivedPrerendersPerPrimaryPageChangedCount
This metric was introduced to investigate the overall usage amount of
speculation rules prerenders in one page. It was estimated by counting
the numbers of started_prerenders_ candidates on PrimaryPageChange.
Usage of started_prerenders_  was reasonable since its candidates are
reset per speculation rules removal, which reflects the developer's
intention.

However, after kPrerender2NewLimitAndScheduler is enabled, non-eager
prerender will be automatically canceled and eliminated from the
started_prerenders_ when reaching the limit. This means that it no
longer hold the prerenders that are expected to be counted when
recording the metric. Additionally, some reset mechanism not limited to
non-eager is planned to be introduced in the near future, so the same
problem could happen.

To solve this, this CL changes the calculation of received prerenders to
simply count up the received prerenders per
PrerendererImpl::MaybePrerender call, which will maintain the original
motivation of the metric and is expected to be consistent with future
changes. Also, this CL adds the "2" to the suffix of the histogram name.

OBSOLETE_HISTOGRAMS=Replaced with
ReceivedPrerendersPerPrimaryPageChangedCount2

Bug: 1464021
Change-Id: I72ae98cb274d94e133492669e494d31fca8fa3ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907848
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Taiyo Mizuhashi <taiyo@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206229}