CCA: Refactor ML service bootstrap in CCA
If the bootstrap of the connection between Ash and ML Service failed,
CCA will hang because CCA is waiting for the callback to be executed
from ML Service but the callback will never be called. This CL
1. Change GetDocumentScannerReadyState with
IsDocumentScannerSupported. The new api just check if CCA supports
document scan mode and doesn't check the loading state.
2. CCA will not wait for document scanner to be ready.
3. Add a disconnect handler to try reconnect to the ML process.

Bug: b:301200244, b:299837974
Test: Manually on DUT and check there is no hang when loading CCA.
Change-Id: I15365583422fcdf73b3451bc859451b770d81466
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4897342
Reviewed-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Sean Li <seannli@google.com>
Cr-Commit-Position: refs/heads/main@{#1206249}