Add UMA for forward migration preceding backward migration
After backward migration is triggered, fetch the stored forward
migration completion time and use it to determine whether forward
migration was performed prior to triggering backward migration or not
(e.g. the user started on Lacros and migration was never performed).

Record the boolean result of this check in UMA.

Add a unit test to verify the new functionality.

Bug: b/245047913
Test: out/Default/unit_tests --gtest_filter=BrowserDataBackMigratorMetricsTest.RecordBackwardMigration*PrecededByForwardMigration
Change-Id: I7c92090f1b1a9438cbc0e9af39bf7c688c8c6809
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916474
Reviewed-by: Victor Gabriel Savu <vsavu@google.com>
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Commit-Queue: Jana Grill <janagrill@google.com>
Cr-Commit-Position: refs/heads/main@{#1206264}