TabDragging tests: Set browser bounds using mouse emulation
On Linux ozone/wayland we can't control the browser bounds, but many
TabDragging tests expect specific bounds and will fail if the actual
bounds are different.

This CL introduces a workaround that allows us to still set the browser
bounds on this platform. The idea is to use ui_controls to emulate mouse
events that will resize and move the browser to where we want (like a
regular user would do).

Bug: 1474921
Change-Id: Ie1e8c154b3f6d078b77e9900f60e7001aa329436
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4738688
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1206297}