Skip backward migration for secondary users
This CL is similar to crrev.com/c/4613565 for forward migration. The
check might not be strictly necessary once Lacros has been completely
disabled for multi user sessions, however it does not break anything so
we should leave it in just in case.

For more details see go/disable-lacros-in-multi-user-session.

Bug: 1454681
Test: out/Default/unit_tests --gtest_filter=BrowserDataBackMigratorShouldMigrateBackTest.*
Change-Id: I93368f3d88ed577f91dcf82155d595d23bb312b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4626216
Reviewed-by: Victor Gabriel Savu <vsavu@google.com>
Commit-Queue: Jana Grill <janagrill@google.com>
Cr-Commit-Position: refs/heads/main@{#1206298}