ImageTransferCache: Apply tone mapping at deserialization
Plumb the HDR headroom from RasterDecoderImpl through to
PaintOp::DeserializeOptions, so that it is available to
PaintOpReader.

Change to the function PaintOpReader::Read(PaintImage*) to now apply
tone mapping at the moment of deserialization. Application of tone
mapping involves creating a new texture with the tone mapping baked
in. This is still inefficient (but significantly better than before).
Add TODOs to move this to playback time.

Remove the application of tone mapping from
ServiceImageTransferCacheEntry::Deserialize.

In GpuImageDecodeCache, do not include the HDR headroom in the
cache key, to avoid re-decoding images every time that the HDR
headroom changes.

Test: OopPixelTest.DrawHdrImageWithMetadata tests this end-to-end

Bug: 1483235
Change-Id: I40c488f8f111ed10c9c7bfd4f0c0bfa05a393110
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4915650
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206306}