PwdMgr: Add comment about kGoogleServicesLastSyncingUsername usage
This pref stores the username (email) of the last user who was
syncing. Key words are "last" (they were syncing before, but might not
be anymore), and "syncing" (only covers users who had Sync-the-feature
enabled).
This CL adds a comment explaining that the "last" part is intentional
(though it might not be strictly required), and the "syncing" part
needs further investigation in the context of upcoming launches.

Bug: 1478765, 1490523
Change-Id: Ibc10744f817615af447c833df615ecd8e936620c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4915939
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206358}