[Omnibox] Copy shortcut_boosted when upgrading match
When suggestions are boosted by the shortcuts provider they can be
ranked with the search suggestion when the feature is enabled.
https://source.chromium.org/chromium/chromium/src/+/main:components/omnibox/browser/autocomplete_result.cc;l=1460-1464;drc=1a511b9d2049a8d4a92a7236288eeca6853220cb

However, the shortcut_boosted property of AutocompleteMatch is not
copied when deduplicating matches. This CL copies the shortcut_boosted
property when deduplicating AutocompleteMatch.

Bug: None
Change-Id: I3085ba78de744c0129110c4a333ac191c79f052c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4915690
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Auto-Submit: Christian Xu <christianxu@chromium.org>
Commit-Queue: Christian Xu <christianxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206390}