Refine BlockReason enum
With this CL we polish the BlockReason enum by renaming the EC Unknown
entry into EC UnknownScanResult which is clearer. In addition, we
include a generic EnterpriseConnectors entry that will be used to
identify EC blocked files with multiple more specific block reasons when
knowing the specific block reason is not required (e.g., no custom
message is defined for that reason).

Bug: b/300611084
Change-Id: Ic748f48f4771e26b61b0b434afdd0e17742b1d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4909642
Commit-Queue: Luca Accorsi <accorsi@google.com>
Reviewed-by: Aida Zolic <aidazolic@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206431}