[soft-navigations] avoid relying on interactionId due to timing issues
A previous CL [1] introduced keyboard event support, and relied on the
interaction count in order to know when to update timestamps. As it
turns out, the interaction count updates are not synchronized with the
time the actual keyboard events fire.
That resulted in race conditions where sometimes all 3 keyboard events
of one interaction plus the first one of the next one all fired before
the interaction count updated. That resulted in a soft navigation
metric report that had the same timestamp as the previous one.

This CL fixes that by moving to use the keydown event as the "barrier"
between interactions.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4906139

Bug: 1489583
Change-Id: I5b951c451b6b1c858bc40085f5f073d2687fb41e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916296
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206434}