[ios] Cleanup BrowserViewWrangler API
Rationalize the API of BrowserViewWrangler to stop exposing internal
logic (such as the requirement to call the methods -createMainBrowser,
-createMainCoordinatorAndInterface and -createInactiveBrowser in that
specific order).

Additionally, the fact that the -incognitoInterface was created in a
lazy way was not used because -[SceneController initialUIMode] needed
to access the -incognitoInterface, thus forcing the creation a few
method call later.

Instead change the BrowserViewWrangler API to

  1. create the three Browser (main, inactive, OTR) in initializer,
  2. create the two interfaces in -createMainCoordinatorAndInterface,
  3. expose a method to request loading the session of all Browsers.

This allow to control when the session is loaded, simplify the logic
to destroy and recreate the OTR ChromeBrowserState (there is no need
to add logic to determine whether the incognito Browser session has
to be loaded or not when the interface is created).

    This is a first step to moving to an asynchronous loading of the
    sessions, but this would require extensive work, so the API is
    left as synchronous for the time being.

Finally, cleanup the implementation of BrowserViewWrangler to remove
internal properties, remove some helper methods that make the code
harder to understand, put methods in the correct `#pragma mark - ...`
section, ...

Update unit tests to reflect the changes made to the API.

Bug: 1383087
Change-Id: Iad85e9b65a3e2db754ebe9dd02203694b33ac2aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4908257
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206437}