[iOS] Repairs AdvancedSettingsSigninTestCase
Those tests fails only on device, not on simulator.
The advanced settings slide up. The tap seems to occur while sliding
up, which caused the tap to miss the button, causing the failures.

Waiting for the app to idle suffice to correct the issue.

Fixed: 1487981
Change-Id: I0fcf6c05cc3ff2895c7c029b0edcb2e9d1ec38b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905407
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206443}