[TTFPassGen] Wire username changed callback in the dailog when saving generated password with empty username
Introduces the custom view (that is the username and password textboxes)
for the modal dialog. It also wires the username changed callback in the
text input control and doesn't save the username in the password store
yet (will be implemented in a separate CL). The password is readonly in
the view and must not be changed.

Bug: 1421753
Change-Id: I2ca198b22f8333fdcbc94680ae3e629b21bf502e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907742
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1206470}