[ios] Large Fakebox focus animation
This change modifies the omnibox focus animation to include
transitioning the height to go from the large fakebox size on the NTP
to the regular omnibox size when focused. When de-focused, the
omnibox animates back to the NTP fakebox size. There is still more
to do to improve the animation, but this handles the size changes.

The animation for non-NTP focus / de-focus is unchanged.

Screen capture of animation:
https://screenshot.googleplex.com/Ba9WMdmRgbAALau.png

Bug: 1482304
Change-Id: I39b1d62397b6aa11159491c97184c6b9ddf953cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4911861
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1206514}