CBD Timeframe Experiment: Add new 15min value to both lists
Add "Last 15 minutes" string to strings and add it to the menu options
that is used in both lists (basic/advanced). It also adds the new value
to the list of options in the dialog once the experiment is on.

Screenshot: https://crbug.com/1487530#c2

Bug: 1487530
Change-Id: I4d28982821b466afc222b47343783540115197f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905297
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Nour Eldien Ayman <noureldien@google.com>
Cr-Commit-Position: refs/heads/main@{#1206523}