Notify OffscreenCanvas of lost context after a GPU crash
In order for an OffscreenCanvas to recover from a lost context caused by
a GPU crash, it has to be notified of it via NotifyGpuContextLost. This
change adds a check before an existing ResourceProvider is returned from
OffscreenCanvas, in order to call NotifyGpuContextLost if the GPU was
being used and its context has been lost.

A test has been added to check that the context is restored correctly
when issuing new draw calls after the GPU crash.

Bug: 1304051
Change-Id: Id6f0df9e0128a62a7d44b044a485f9b9e83b5d4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4854158
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Andres Ricardo Perez Rojas <andresrperez@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206543}