view-transition: Fix DOMViewTransition null-ptr crash
InvokeDOMUpdate can be called asynchronously from SkipTransition. If the
owning frame is disconnected before the task is executed, calling
ToScriptState on its execution_context_ will result in a nullptr.
Early-return to avoid crashing. There's no need to fulfill the promises
in this case, or reply to the ViewTransition object, since the Document
is shutting down in this case anyway.

Bug: 1489642
Change-Id: I19d213831367957bd12b719edcab07a5e11f941f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916356
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206552}