Add a FrameContext test without PerformanceManagerTestHarness.
This is a more complete test of the case where PM doesn't know about a
RenderFrameHost, since it checks what happens if a user calls FrameContext::FromRenderFrameHost() before PM is initialized.

Also moves RunInGraph to a helper function so it can be used outside
the test harness.

Bug: 1471683
Change-Id: I9633e3875ef593f04db0bd3aa205501642a86d83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916440
Commit-Queue: Joe Mason <joenotcharles@google.com>
Reviewed-by: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206565}