[iOS] Fix DCHECK Hit in Privacy Safe Browsing Settings Page
The DCHECK was getting hit due to an asynchronized UI sequence that
happens when creating coordinators. This happens when a user opens up
the setting page, slides down to remove it, and then quickly opens the
setting page again. Although this is a rare occurrence, DCHECKs are
supposed to be invariant, and seeing as this is possible, we should
do any clean up explicitly to ensure that a coordinator is reset
when re-opening it.

Fixed: 1489105
Change-Id: Ib3a0ab30a6da12385165b5baa7a677cc53c9c355
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916538
Reviewed-by: Ali Juma <ajuma@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206604}