[PolicyLogs] Adding Log for EnableChromeBrowserCloudManagement Flag
This CL adds text to a policy log to ask user to enable the command line switch/flag if it is not set in a non chrome-branded build.
`IsEnabled()` always returns true in branded builds, so this log is mainly for debugging purposes.

A duplicate log is also removed to reduce log clutter.


Change-Id: Ia8e7457ecb242e7f61f3dc339394bc0dc2b11615
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916362
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1206621}