[wallets] Refactor: Use Digital Credential as a generic term for MDoc
This CL is a pure rename refactoring from MDocs -> Digital Credential,
inspired by the choice of terms that the OpenID4VP group has taken:

https://openid.net/wg/digital-credentials-protocols/

We made the API format agnostic, so we can use the generic term
Digital Credential (in the context of the digitalization of government
issued credentials).

Verifiable Credentials occurred to us as another
generic option, but we were worried it could be confused with VC, the
spec.

Bug: 1416939
Change-Id: Id8a8396e0064747010c8bd85c773a49ebdd0bbac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4917013
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Sam Goto <goto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206631}