[SidePanelPinning] Move the browser root action to browser instead of browser view.
Previously the the root action was tied to the BrowserView. However, since many views are created in the construction of the BrowserView the Browser does not have access to the window(). This means that the views cannot translate a browser to a BrowserView at that time in their initialization.

This CL adds the root_action_item to the Browser. Since BrowserView holds an unique_ptr to this, it sets the actionItem in its action_initializer.

Change-Id: I2b88a5ea5a5cc16d7d8c32fbf9e58b4a4e45b01b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4915476
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Shibalik Mohapatra <shibalik@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1206632}