Update ContentAnalysisDownloadsDelegate to observe its download
ContentAnalysisDownloadsDelegate holds a raw_ptr to its corresponding
DownloadItem, which can be dangling. This leads to a crash in dev
builds since the download is destroyed before the delegate, so a fix
to prevent this and future bugs if the logic in
ContentAnalysisDownloadsDelegate becomes more intricate is to make
it observe the download and update its pointer when that download is
destroyed.

Change-Id: I0211c988be4390f815756f693ae82f966303e2af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4919435
Reviewed-by: Nasser Al-shawwa <alshawwa@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Auto-Submit: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206645}